+++ src/rename-angular-component/in-file-edits/strip-selector-braces.function.ts
@@ -1,0 +1,6 @@
+export function stripSelectorBraces(selector: string, andQuotes = false) {
+  if (andQuotes) {
+    selector = selector.replace(/'/g, '');
+  }
+  return selector.replace(/(\[|\]|\.)/g, '');
+}