+++ src/test/suite/helpers/constants-helper-config.ts
@@ -1,0 +1,3 @@
+export const OVERWRITE_DIFF_SNAPSHOTS = false; // false
+
+export const DISCARD_STAGED_CHANGES = true; // true