+++ src/test/suite/scenario1.test.ts
@@ -5,6 +5,11 @@ suite('Suite Scenario 1', () => {
   vscode.window.showInformationMessage('Start all tests.');
 
   test('Scenario 1', async () => {
+    /*
+    This test scenario uses publicly available repo: https://github.com/tomwhite007/simple-reactive-viewmodel-example
+    as the seed to base 1 renamer test on.
+    */
+
     await genericTestScenario({
       projectRoot:
         '/Users/tom/Development/vscode-ext/_rename-test-spas/simple-reactive-viewmodel-example',