+++ src/test/suite/scenario4.test.ts
@@ -5,6 +5,11 @@ suite('Suite Scenario 4', () => {
   vscode.window.showInformationMessage('Start all tests.');
 
   test('Scenario 4', async () => {
+    /*
+    This test scenario uses a private repo as the seed to base 1 renamer test on.
+    See src/test/suite/scenario1.test.ts and src/test/suite/scenario5.test.ts for publicly available test repos.
+    */
+
     await genericTestScenario({
       projectRoot:
         '/Users/tom/Development/vscode-ext/_rename-test-spas/test-paths-app',